code and docs refactor
