+++ .gitignore
@@ -4,3 +4,4 @@ exdata/
 optimizers/__pycache__
 ppdata/
 venv/
+*.pth
